Bugs fixed

+++ rs.h
@@ -69,7 +69,7 @@ typedef __m128i *contr;
 typedef uint32_t meter;
 typedef uint16_t place;
 typedef uint16_t agent;
-typedef uint16_t penny;
+typedef uint32_t penny;
 typedef uint16_t edge;
 typedef uint32_t id;
 typedef float dist;
